// client/src/app/dataModels/catalog/productDetails/binocularDetails.ts
import { BinocularPurpose } from "../../enums/binocuclar/binocularPurpose";
import { FocusingMethod } from "../../enums/binocuclar/focusingMethod";
import { OpticsMaterial } from "../../enums/binocuclar/opticsMaterial";
import { PrismType } from "../../enums/binocuclar/prismType";
import { ProductCharacteristics } from "./productCharacteristics";
import { ProductDetails } from "./productDetails";

export interface BinocularDetails extends ProductDetails, ProductCharacteristics {
    aperture?: number;
    exitPupilDiameterMax?: number;
    exitPupilDiameterMin?: number;
    focusingMethod?: FocusingMethod;
    fovMin?: number;
    fovMax?: number;
    hasAdapter?: string;
    hasCase?: string;
    hasMoistureProtection?: string;
    interpupillaryDistanseMin?: number;
    interpupillaryDistanseMax?: number;
    focusDistanceMin?: number;
    opticsMaterial?: OpticsMaterial;
    prismType?: PrismType;
    binocularPurpose?: BinocularPurpose;
    relativeBrightnessMin?: number;
    relativeBrightnessMax?: number;
    removalExitPupilMin?: number;
    removalExitPupilMax?: number;
    scaleMin?: number;
    scaleMax?: number;
    weight?: number;
}